==> notebooks/data_processing/01_data_aquisition.ipynb <==
# **Data Aquisition Notebook**

This notebook aims to centralize the code that captures the data necessary for the development of the trading advisor project. We will use data sources such as Nasdaq Data Link, Yahoo Finance, and Investpy.

## **Initial Setup**

This initial setup section is responsible for managing and installing the necessary packages for running the notebook that acquires the data used in the project. It's worth noting that I also provide the requirements.txt file where all the packages are centralized, and it can be executed in a more silent manner.

### Install Libs

%pip install --upgrade pip --q --no-cache
%pip install pandas --q --no-cache
%pip install python-dotenv --q --no-cache
%pip install yfinance --q --no-cache
%pip install Nasdaq-Data-Link --q --no-cache
%pip install investpy --q --no-cache

### Import Libs

import os
from pathlib import Path
import pandas as pd
from dotenv import load_dotenv
import nasdaqdatalink as ndl
import yfinance as yf
import investpy as inv

### Load enviroments var

load_dotenv()

### Fullfilment constants vars

NASDAQ_API_KEY = os.getenv("NASDAQ_API_KEY")
ndl.ApiConfig.api_key = NASDAQ_API_KEY

country_name = "brazil"
start_date = "2019-01-01"

### Create a file path default

file_path_raw = str(Path(os.getcwd()).parent/"data/raw")

## **Collecting Data**

This data collection section is responsible for retrieving data from the sources for the respective processing. We will collect fundamental and balance sheet data for publicly listed companies, historical data for key macroeconomic indices, and, finally, historical stock price data for the companies' shares.

### Fundamentals

* Listing all publicly traded companies in Brazil.

df_compaines = inv.stocks.get_stocks(country = country_name)
df_compaines.head(2)

* Creating a list with the company symbols and adding the (.SA) characters. Adding these characters is necessary as it's required to explicitly specify to the Yahoo Finance API that these companies are from South America (SA).

tickers = list(df_compaines["symbol"])
tickers_filtered = [tickers_filtered + ".SA" for tickers_filtered in tickers if len(tickers_filtered) <= 5]

* Creating the get_fundamentals_data function to make it feasible to extract data from the Yahoo API. Notice that we've created a dictionary with the names of the main columns we will use. It's worth noting that the Yahoo API returns numerous columns that wouldn't be used in the project.

def get_fundamentals_data(tickers):

    df_list = []

    for ticker in tickers:
        try:
            company = yf.Ticker(ticker)
            fundamental_data = company.info

            fundamental_data = {
                'ticker': fundamental_data.get('symbol'),
                'long_name': fundamental_data.get('longName'),
                'sector': fundamental_data.get('sector'),
                'industry': fundamental_data.get('industry'),
                'market_cap': fundamental_data.get('marketCap'),
                'enterprise_value': fundamental_data.get('enterpriseValue'),
                'total_revenue': fundamental_data.get('totalRevenue'),
                'profit_margins': fundamental_data.get('profitMargins'),
                'operating_margins': fundamental_data.get('operatingMargins'),
                'net_income': fundamental_data.get('netIncome'),
                'dividend_rate': fundamental_data.get('dividendRate'),
                'beta': fundamental_data.get('beta'),
                'ebitda': fundamental_data.get('ebitda'),
                'trailing_pe': fundamental_data.get('trailingPE'),
                'forward_pe': fundamental_data.get('forwardPE'),
                'volume': fundamental_data.get('volume'),
                'average_volume': fundamental_data.get('averageVolume'),
                'fifty_two_week_low': fundamental_data.get('fiftyTwoWeekLow'),
                'fifty_two_week_high': fundamental_data.get('fiftyTwoWeekHigh'),
                'price_to_sales_trailing_12_months': fundamental_data.get('priceToSalesTrailing12Months'),
                'fifty_day_average': fundamental_data.get('fiftyDayAverage'),
                'two_hundred_day_average': fundamental_data.get('twoHundredDayAverage'),
                'trailing_annual_dividend_rate': fundamental_data.get('trailingAnnualDividendRate'),
                'trailing_annual_dividend_yield': fundamental_data.get('trailingAnnualDividendYield'),
                'book_value': fundamental_data.get('bookValue'),
                'price_to_book': fundamental_data.get('priceToBook'),
                'total_cash': fundamental_data.get('totalCash'),
                'total_cash_per_share': fundamental_data.get('totalCashPerShare'),
                'total_debt': fundamental_data.get('totalDebt'),
                'earnings_quarterly_growth': fundamental_data.get('earningsQuarterlyGrowth'),
                'revenue_growth': fundamental_data.get('revenueGrowth'),
                'gross_margins': fundamental_data.get('grossMargins'),
                'ebitda_margins': fundamental_data.get('ebitdaMargins'),
                'operating_margins': fundamental_data.get('operatingMargins'),
                'return_on_assets': fundamental_data.get('returnOnAssets'),
                'return_on_equity': fundamental_data.get('returnOnEquity'),
                'gross_profits': fundamental_data.get('grossProfits')
            }

            df_ticker = pd.DataFrame([fundamental_data])
            df_list.append(df_ticker)

        except Exception as e:
            print(f"Error processing ticker {ticker}: {e}")

    if df_list:
        df_fundamentals = pd.concat(df_list, ignore_index=True)
    else:
        df_fundamentals = pd.DataFrame()

    return df_fundamentals

df_fundamentals = get_fundamentals_data(tickers_filtered)

* When executing the get_fundamentals_data function, we encounter exceptions in data retrieval for some companies, for example (Error processing ticker JBDU1.SA: 404 Client Error: Not Found for url: ...). This error occurs because the Yahoo API couldn't find data for the company, either because the company is no longer listed on the stock exchange or has changed its symbol, as in the case of Via Varejo, which has changed its symbol several times (VVIA3, VIIA3, and BHIA3).

df_fundamentals.head(2)

Path(file_path_raw).mkdir(parents=True, exist_ok=True)
df_fundamentals.to_csv(file_path_raw + "/fundamentals_raw.csv", index=False)

### Stocks

* The stocks section is responsible for collecting historical price data of stocks traded on the stock exchange.

def get_stocks_data(tickers_filtered, start_date):

    df = yf.download(tickers_filtered,start=start_date)
    _stacked = df.stack()
    _stacked.reset_index(inplace=True)
    _stacked.rename(columns={"level_1" : "ticker"}, inplace=True)
    df_stocks = _stacked

    return df_stocks

df_stocks = get_stocks_data(tickers_filtered, start_date)

* When executing the get_stocks_data function, we encounter a progress bar that indicates the completion percentage and the actions that couldn't be collected for reasons already known. These reasons may include the company changing its ticker symbol or no longer being traded on the stock exchange. In the case of stocks, it may also be that the company is not traded on the spot market

df_stocks.head(2)

Path(file_path_raw).mkdir(parents=True, exist_ok=True)
df_stocks.to_csv(file_path_raw + "/stocks_raw.csv", index=False)

### Macroeconomic

* TODO - FIX 404 NasdaqDataLink

* The macroeconomics section is responsible for collecting the country's key macroeconomic data, such as IPCA, GDP, SELIC, and others.

def get_marcoeconomic_data(start_date):

    df_selic = pd.DataFrame()
    df_consumer_confidence = pd.DataFrame()
    df_pib = pd.DataFrame()
    df_incc = pd.DataFrame()
    df_ipca = pd.DataFrame()
    df_dolar = pd.DataFrame()

    df_selic['selic'] = ndl.get('BCB/432', start_date = start_date, collapse = "monthly")
    df_consumer_confidence['confidence'] = ndl.get('BCB/4393', start_date = start_date, collapse = "monthly")
    df_pib['pib'] = ndl.get("BCB/4380", start_date = start_date, collapse = "monthly")
    df_incc['incc'] = ndl.get('BCB/192', start_date = start_date, collapse = "monthly")
    df_ipca['ipca'] = ndl.get("BCB/13522", start_date = start_date, collapse = "monthly")
    df_dolar['dolar'] = ndl.get('BCB/10813', start_date = start_date, collapse = "monthly")

    df_macroeconomic = pd.concat([df_selic, df_consumer_confidence, df_pib, df_incc, df_ipca, df_dolar], axis=1)
    df_macroeconomic = df_macroeconomic.reset_index()
    
    return df_macroeconomic

df_macroeconomic = get_marcoeconomic_data(start_date)

df_macroeconomic.head(2)

Path(file_path_raw).mkdir(parents=True, exist_ok=True)
df_macroeconomic.to_csv(file_path_raw + "/macroeconomic_raw.csv", index=False)

* We can identify that at the end of the generated dataset, there are null values in the last two rows. This occurs due to the regulatory body's meeting frequency for these indices, which occurs every 45 days.
