==> notebooks/data_processing/02_data_cleaning.ipynb <==
# **Data Cleaning Notebook**

This notebook's central mission is to consolidate data cleaning tasks, such as removing columns and rows with missing values, renaming columns to adhere to PEP8 best practices, and more. It's essential to note that the data in question was captured by the "01_data_aquisition.ipynb" notebook. Therefore, it is of paramount importance to ensure that the "01_data_aquisition.ipynb" notebook has been properly executed before running this one.

## **Initial Setup**

This initial setup section is responsible for managing and installing the necessary packages for running the notebook that acquires the data used in the project. It's worth noting that I also provide the requirements.txt file where all the packages are centralized, and it can be executed in a more silent manner.

### Import Libs

import os
import pandas as pd
from pathlib import Path

### Create a file path default

#### Full Data

file_path_raw = str(Path(os.getcwd()).parent/"data/raw")
file_path_cleaned  = str(Path(os.getcwd()).parent/"data/cleaned")

## **Cleaning Data**

* In this section, the data cleaning process is presented, which includes tasks such as filling or removing rows and columns with missing values, renaming, and retaining the necessary columns for the project's execution.

### Fundamentals Data

df_fundamentals_raw = pd.read_csv(file_path_raw + "/fundamentals_raw.csv")
df_fundamentals_raw.head()

#### Cleaning the columns

df_fundamentals_raw.columns

df_fundamentals_cols = df_fundamentals_raw.drop(columns=["net_income"]).copy()
df_fundamentals_cols

* In the cell above, the "net_income" columns were dropped because all the values were null, and the "Unnamed: 0" column was removed, which appears due to the leading comma (,) in the CSV file when a DataFrame is saved.

#### Cleaning the missing values

df_fundamentals_cols.isna().sum()

df_fundamentals_missing = df_fundamentals_cols.dropna(subset=['long_name', 'sector', "industry", "market_cap", "enterprise_value", "total_revenue", "profit_margins", "operating_margins", "beta"]).copy()
df_fundamentals_missing.isna().sum()

* In the cell above, all rows containing null values in the columns ('long_name', 'sector', "industry", "market_cap", "enterprise_value", "total_revenue", "profit_margins", "operating_margins", "beta") were removed because it was already known that in the data collection, there were companies that were no longer being traded on the stock exchange or had changed their tickers. Given the importance of these mentioned columns, it is prudent not to accept any null values in these fields, as these values are believed to be of utmost importance for future analysis, and filling them with computed data or any business rule is too risky.

df_fundamentals_missing.fillna(0, inplace=True)
df_fundamentals_missing.isna().sum()

* In the cell above, the rest of the columns with missing values were filled with the number 0. After some analysis, it was understood that these missing values occur because either the company does not engage in the activity of paying dividends, as in the case of the "dividend_rate" field, or the company did not disclose certain results and indicators. It is worth noting that any data manipulation to fill a missing value should be done with great care.

df_fundamentals_cleaned = df_fundamentals_missing.copy()

df_fundamentals_raw.shape

df_fundamentals_cleaned.shape

#### Saving files

Path(file_path_cleaned).mkdir(parents=True, exist_ok=True)
df_fundamentals_cleaned.to_csv(file_path_cleaned + "/fundamentals_cleaned.csv", index=False)

### Macroeconomic Data

df_macroeconomic_raw = pd.read_csv(file_path_raw + "/macroeconomic_raw.csv")
df_macroeconomic_raw.head()

#### Cleaning the Columns

df_macroeconomic_cols = df_macroeconomic_raw.rename(columns={"Date" : "date"}).copy()
df_macroeconomic_cols.columns

#### Cleaning the missing values

df_macroeconomic_cols.isna().sum()

df_macroeconomic_missing = df_macroeconomic_cols.dropna().copy()
df_macroeconomic_missing.isna().sum()

* In the macroeconomics dataframe, null values were removed because, as known, there is a meeting every 45 days with the Copom for the new definitions of these indices.

df_macroeconomic_cleaned = df_macroeconomic_missing.copy()

df_macroeconomic_raw.shape

df_macroeconomic_cleaned.shape


#### Saving files

Path(file_path_cleaned).mkdir(parents=True, exist_ok=True)
df_macroeconomic_cleaned.to_csv(file_path_cleaned + "/macroeconomic_cleaned.csv", index=False)

### Stocks Data

df_stocks_raw = pd.read_csv(file_path_raw + "/stocks_raw.csv")
df_stocks_raw.head()

#### Cleaning the Columns

df_stocks_cols = df_stocks_raw.rename(columns=lambda col: col.lower().replace(' ', '_')).copy()
df_stocks_cols.columns


#### Clearning the missing values

df_stocks_missing = df_stocks_cols.copy()
df_stocks_missing.isna().sum()

* In the case of stock data, it wasn't necessary to exert much effort because the data came in a clean format, largely due to the care that the Yahoo Finance API gives to this information.

df_stocks_cleaned = df_stocks_missing.copy()

df_stocks_raw.shape

df_stocks_cleaned.shape

#### Saving files

Path(file_path_cleaned).mkdir(parents=True, exist_ok=True)
df_stocks_cleaned.to_csv(file_path_cleaned + "/stocks_cleaned.csv", index=False)
