==> notebooks/data_processing/03_data_book.ipynb <==
# **Data Book Notebook**

* This section aims to enrich and centralize the necessary data for the project's execution. Some of this data will be created through calculations based on existing data.

## **Initial Setup**

### Install Libs

%pip install scikit-learn --q --no-cache

### Import Packages

import os
from pathlib import Path
import pandas as pd
import numpy as np
from sklearn.impute import KNNImputer

### Create a file path default

file_path_cleaned = str(Path(os.getcwd()).parent/"data/cleaned")
file_path_book  = str(Path(os.getcwd()).parent/"data/book")

## **Data Book**

* This section's mission is to enrich the data captured and cleaned by the processes in the notebooks "01_data_aquisition.ipynb" and "02_data_cleaning.ipynb". This enrichment will be achieved through calculations with existing fields and possibly the acquisition of new external data. It's important to emphasize the significance of running the previous notebooks for the proper functioning of the project.

### Fundamentals Book

df_fundamentals_cleaned = pd.read_csv(file_path_cleaned + "/fundamentals_cleaned.csv")
df_fundamentals_cleaned.head()

#### Feature Creation

df_fundamentals_book_cols = df_fundamentals_cleaned.copy()

df_fundamentals_book_cols['total_assets_approx'] = df_fundamentals_book_cols['total_cash'] + df_fundamentals_book_cols['book_value']
df_fundamentals_book_cols['asset_turnover'] = df_fundamentals_book_cols['total_revenue'] / df_fundamentals_book_cols['total_assets_approx']
df_fundamentals_book_cols['earnings_growth_rate'] = df_fundamentals_cleaned['earnings_quarterly_growth'] * 100
df_fundamentals_book_cols['dividend_payout_ratio'] = (df_fundamentals_cleaned['trailing_annual_dividend_rate'] / df_fundamentals_cleaned['earnings_quarterly_growth']) * 100
df_fundamentals_book_cols['dividend_payout_ratio'] = df_fundamentals_book_cols['dividend_payout_ratio'].replace([np.inf, -np.inf], np.nan)
df_fundamentals_book_cols['dividend_payout_ratio'] = df_fundamentals_book_cols['dividend_payout_ratio'].apply(lambda x: max(0, x))
df_fundamentals_book_cols['equity'] = df_fundamentals_book_cols['total_assets_approx'] - df_fundamentals_book_cols['total_debt']
df_fundamentals_book_cols['debt_to_equity'] = df_fundamentals_book_cols['total_debt'] / df_fundamentals_book_cols['equity']
df_fundamentals_book_cols['roi'] = df_fundamentals_cleaned['total_revenue'] / df_fundamentals_cleaned['enterprise_value']
df_fundamentals_book_cols['roce'] = df_fundamentals_book_cols['ebitda'] / (df_fundamentals_book_cols['total_debt'] + df_fundamentals_book_cols['market_cap'])

* In the code block above, we are working with a DataFrame of company fundamentals. The fields "earnings_growth_rate," "dividend_payout_ratio," and "roi" play pivotal roles in our analysis. "Earnings_growth_rate" informs us about the company's profit growth over time, which is crucial for assessing its potential for appreciation. Meanwhile, "dividend_payout_ratio" aids us in understanding the company's dividend policy, revealing how much of the profits is distributed to shareholders. Lastly, "ROI" is an essential metric for evaluating financial efficiency and company performance in relation to invested capital. These fields are the cornerstones of our analysis and guide our investment decisions.

df_fundamentals_book_cols.columns

df_fundamentals_book_cols.head(2)

#### Cleaning Missing Values

df_fundamentals_book_missing = df_fundamentals_book_cols.copy()

df_fundamentals_book_missing.isna().sum()

df_fundamentals_book_missing["dividend_payout_ratio"].fillna(0, inplace = True)

df_fundamentals_book_missing.isna().sum()

df_fundamentals_book = df_fundamentals_book_missing.copy()

df_fundamentals_book.head()

#### Saving Files

Path(file_path_book).mkdir(parents=True, exist_ok=True)
df_fundamentals_book.to_csv(file_path_book + "/fundamentals_book.csv", index=False)

### Macroeconomic Book

df_macroeconomic_cleaned = pd.read_csv(file_path_cleaned + "/macroeconomic_cleaned.csv")
df_macroeconomic_cleaned.head()

#### Feature Creation

df_macroeconomic_book_cols = df_macroeconomic_cleaned.copy()

df_macroeconomic_book_cols['monthly_inflation'] = df_macroeconomic_book_cols['ipca'].pct_change()
df_macroeconomic_book_cols['gdp_growth'] = df_macroeconomic_book_cols['pib'].pct_change()
df_macroeconomic_book_cols['dollar_growth'] = df_macroeconomic_book_cols['dolar'].pct_change()
df_macroeconomic_book_cols['real_interest_rate'] = df_macroeconomic_book_cols['selic'] - df_macroeconomic_book_cols['ipca']
df_macroeconomic_book_cols['inflation_confidence_difference'] = df_macroeconomic_book_cols['ipca'] - df_macroeconomic_book_cols['confidence']

df_macroeconomic_book_cols.columns

* In the code block above, we are working with a DataFrame of economic indicators. The fields "monthly_inflation," "gdp_growth," "dollar_growth," "real_interest_rate," and "inflation_confidence_difference" are pivotal in our economic analysis. "Monthly_inflation" informs us about the monthly changes in inflation, which is vital for understanding price stability. "GDP_growth" tracks the growth rate of the Gross Domestic Product, indicating the economic health and expansion of the country. "Dollar_growth" reflects changes in the value of the national currency relative to the US dollar, providing insights into exchange rate fluctuations. "Real_interest_rate" measures the interest rate adjusted for inflation, helping us assess borrowing costs and investment opportunities. Lastly, "Inflation_confidence_difference" quantifies the difference in inflation expectations, shedding light on economic confidence levels. These fields are integral to our economic analysis and inform our decisions.

df_macroeconomic_book_cols.head(5)

#### Cleaning MIssing Values

df_macroeconomic_book_missing = df_macroeconomic_book_cols.copy()

df_macroeconomic_book_missing.isna().sum()

dates = df_macroeconomic_book_missing['date']
numerical_columns = ['selic', 'confidence', 'pib', 'incc', 'ipca', 'dolar', 'monthly_inflation', 'gdp_growth', 'dollar_growth', 'real_interest_rate', 'inflation_confidence_difference']

imputer = KNNImputer(n_neighbors=5)
imputed_values = imputer.fit_transform(df_macroeconomic_book_missing[numerical_columns])
df_macroeconomic_book_missing = pd.DataFrame(imputed_values, columns=numerical_columns)
df_macroeconomic_book_missing['date'] = dates
df_macroeconomic_book_missing = df_macroeconomic_book_missing[['date','selic', 'confidence', 'pib', 'incc', 'ipca', 'dolar', 'monthly_inflation', 'gdp_growth', 'dollar_growth', 'real_interest_rate', 'inflation_confidence_difference']]

* In the code above, missing data in the macroeconomic dataset stored in the "df_macroeconomic_book_missing" DataFrame was addressed. The process involved the use of the KNNImputer, an imputation method based on the k-nearest neighbors algorithm. With a parameter of n_neighbors set to 5, it considered the five nearest neighbors for each missing value in the numerical columns. The resulting imputed values are stored in the "df_macroeconomic_book_missing" DataFrame. Additionally, a 'date' column was appended to facilitate time-based analysis, ensuring that imputed values are associated with their respective dates.

df_macroeconomic_book_missing.isna().sum()

#### Save Files

df_macroeconomic_book = df_macroeconomic_book_missing.copy()

Path(file_path_book).mkdir(parents=True, exist_ok=True)
df_macroeconomic_book.to_csv(file_path_book + "/macroeconomic_book.csv", index=False)

### Stocks Book

df_stocks_cleaned = pd.read_csv(file_path_cleaned + "/stocks_cleaned.csv")
df_stocks_cleaned.head()

#### Feature Creation

df_stocks_book_cols = df_stocks_cleaned.copy()

df_stocks_book_cols['daily_return'] = df_stocks_book_cols['adj_close'].pct_change()
df_stocks_book_cols['short_term_moving_average'] = df_stocks_book_cols['adj_close'].rolling(window=10).mean()
df_stocks_book_cols['moving_average_mean'] = df_stocks_book_cols['adj_close'].rolling(window=50).mean()
df_stocks_book_cols['long_term_moving_average'] = df_stocks_book_cols['adj_close'].rolling(window=200).mean()

window = 14
delta = df_stocks_book_cols['adj_close'].diff()
gain = delta.where(delta > 0, 0)
loss = -delta.where(delta < 0, 0)
average_gain = gain.rolling(window=window).mean()
average_loss = loss.rolling(window=window).mean()
rs = average_gain / average_loss
df_stocks_book_cols['rsi'] = 100 - (100 / (1 + rs))

df_stocks_book_cols['volatility'] = df_stocks_book_cols['adj_close'].pct_change().rolling(window=20).std()
df_stocks_book_cols['average_volume'] = df_stocks_book_cols['volume'].rolling(window=20).mean()
df_stocks_book_cols['daily_range'] = df_stocks_book_cols['high'] - df_stocks_book_cols['low']
df_stocks_book_cols['open_close_difference'] = df_stocks_book_cols['open'] - df_stocks_book_cols['adj_close']
df_stocks_book_cols['high_low_difference'] = df_stocks_book_cols['high'] - df_stocks_book_cols['low']
df_stocks_book_cols['acceleration'] = df_stocks_book_cols['adj_close'].pct_change().rolling(window=5).mean()
df_stocks_book_cols['hammer'] = (df_stocks_book_cols['open'] > df_stocks_book_cols['adj_close']) & (df_stocks_book_cols['low'] > df_stocks_book_cols['adj_close'] - 0.01)
df_stocks_book_cols['cagr'] = ((df_stocks_book_cols['adj_close'] / df_stocks_book_cols['adj_close'].shift(1)) - 1).mean()


In this section, we explore crucial financial metrics in our dataset:

- **daily_return**: Tracks the daily percentage change in asset prices, revealing daily performance dynamics.

- **short_term_moving_average**: Computes the average asset price over a brief time frame, aiding short-term trend analysis.

- **moving_average_mean**: Highlights the mean price over a specified time window, offering insights into overall price trends.

- **long_term_moving_average**: Reflects the average asset price over an extended duration, facilitating long-term trend assessment.

- **rsi (Relative Strength Index)**: Measures price momentum, aiding in identifying overbought or oversold conditions.

- **volatility**: Quantifies price variability, assisting in risk evaluation and market analysis.

- **average_volume**: Indicates average trading volume, valuable for assessing asset liquidity and investor interest.

- **daily_range**: Measures the price range between daily highs and lows, a critical factor for risk evaluation.

- **open_close_difference**: Calculates the difference between opening and closing prices, revealing intraday price movements.

- **high_low_difference**: Quantifies the spread between daily high and low prices, aiding in identifying price extremes.

- **acceleration**: Analyzes the rate of price change, providing insights into price movement speed.

- **hammer**: Identifies hammer candlestick patterns for potential trend reversals in technical analysis.

- **cagr (Compound Annual Growth Rate)**: Computes the annual growth rate of an investment over a specified period, valuable for assessing long-term performance.


df_stocks_book_cols.head(2)

#### Cleaning Missing Values

df_stocks_book_missing = df_stocks_book_cols.copy()

df_stocks_book_missing.isna().sum()

numeral_columns = df_stocks_book_missing.select_dtypes(include=['number'])
numeral_values = df_stocks_book_missing[numeral_columns.columns]

imputer = KNNImputer(n_neighbors=5)
imputed_values = imputer.fit_transform(numeral_values)
numeral_values = pd.DataFrame(imputed_values, columns=numeral_values.columns)

categorias_columns = df_stocks_book_missing.select_dtypes(exclude=['number'])
categorias_values = df_stocks_book_missing[categorias_columns.columns]

df_stocks_book_missing = pd.concat([numeral_values, categorias_values], axis=1)
df_stocks_book_missing.head(2)

df_stocks_book_missing.isna().sum()

* In the code above, missing data in the macroeconomic dataset stored in the "df_macroeconomic_book_missing" DataFrame was addressed. The process involved the use of the KNNImputer, an imputation method based on the k-nearest neighbors algorithm. With a parameter of n_neighbors set to 5, it considered the five nearest neighbors for each missing value in the numerical columns. The resulting imputed values are stored in the "df_stocks_book_missing" DataFrame. Additionally, a 'date' column was appended to facilitate time-based analysis, ensuring that imputed values are associated with their respective dates.

df_stocks_book = df_stocks_book_missing.copy()
df_stocks_book = df_stocks_book[['ticker','adj_close', 'close', 'high', 'low', 'open', 'volume', 'daily_return', 'short_term_moving_average', 'moving_average_mean', 'long_term_moving_average', 'rsi', 'volatility', 'average_volume', 'daily_range', 'open_close_difference', 'high_low_difference', 'acceleration', 'cagr', 'date', 'hammer']]

##### Save Files

Path(file_path_book).mkdir(parents=True, exist_ok=True)
df_stocks_book.to_csv(file_path_book + "/stocks_book.csv", index=False)
